% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessGWAS.R
\name{get.regions}
\alias{get.regions}
\title{Define regions aroung distinct association signals from all traits of interest}
\usage{
get.regions(filename.lst, path, wd = 1e+06)
}
\arguments{
\item{path}{filepath to save regions for trait}

\item{wd}{window around signals, defines regions' size}

\item{file.lst}{list of paths for excel file containing indpendent signals for the traits of interest (columns: SNPID, CHR, POS)}
}
\value{
regions to perform colocalization
}
\description{
Define regions aroung distinct association signals from all traits of interest
}
