% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Welcome user}
\usage{
hello(x)
}
\arguments{
\item{x}{user's name}
}
\value{
The output from \code{\link{print}}
}
\description{
Welcome user
}
\examples{
hello("Ana")
\dontrun{hello("Ole")}
}
