% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{lookup.DEG}
\alias{lookup.DEG}
\title{Look up genes in DEGs}
\usage{
lookup.DEG(
  DEG,
  genes,
  file.path = "lookups/DEG.csv",
  logFC.theta = log(1.5, base = 2),
  alpha = 0.05
)
}
\arguments{
\item{DEG}{named list of data tables containing DEGs for a trait (columns: gene, ID, logFC, pval.adj)}

\item{genes}{data table of genes of interest (columns: name, ID)}

\item{file.path}{path to save result table, default="lookups/DEG.csv"}

\item{logFC.theta}{threshold for abs(logFC), default=log(1.5, base=2)}

\item{alpha}{significance threshold for p-value, default=0.05}
}
\value{
a data table of genes with 1 or 0 in each DEG column
}
\description{
Check if any gene in your list is a DEG for any trait
}
