% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{lookup.hc}
\alias{lookup.hc}
\title{Look up genes in high confidence genes data}
\usage{
lookup.hc(hc.genes, genes, file.path = "lookups/HC.csv")
}
\arguments{
\item{genes}{data table of genes of interest (columns: name, ID)}

\item{file.path}{path to save result table, default="lookups/HC.csv"}

\item{hc}{named list of high confidence genes for each trait of interest}
}
\value{
a data table of genes with 1 or 0 for each trait column
}
\description{
Check if any gene in your list is already a high confidence likely effector gene for any of the traits of interest
}
