% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{lookup.OMIM}
\alias{lookup.OMIM}
\title{Look up genes in OMIM}
\usage{
lookup.OMIM(terms.lst, genes, file.path = "lookups/OMIM.csv")
}
\arguments{
\item{genes}{data table of genes of interest (columns: name, ID)}

\item{file.path}{path to save result table, default="lookups/OMIM.csv"}

\item{terms}{named list of terms related to each trait of interest}
}
\value{
a data table of genes with 1 or 0 for each trait column
}
\description{
Look up genes in OMIM
}
