% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessGWAS.R
\name{read_GWAS}
\alias{read_GWAS}
\title{Define regions aroung distinct association signals from all traits of interest}
\usage{
read_GWAS(gwas.file, dict)
}
\arguments{
\item{gwas.file}{excluding indels (columns: CHR, POS, rsID, pval, beta, se, MAF, EAF, N, Ncases, EA, NEA)}

\item{wd}{window around signals, defines regions' size}
}
\value{
regions to perform colocalization
}
\description{
Define regions aroung distinct association signals from all traits of interest
}
