% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{search.OMIM}
\alias{search.OMIM}
\title{Search OMIM}
\usage{
search.OMIM(gene)
}
\arguments{
\item{gene}{gene symbol}
}
\value{
data table of OMIM's entry for the gene
}
\description{
Get OMIM's entry for a gene
}
